# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateNodeRequest": {
        "description": "节点注册请求：兼容两种格式\n- 面板/测试：`{\"node\": {name, hostname, ip_address, status, ...}}`\n- Agent 平铺：`{name, host, agent_port, auth_token}`",
        "properties": {
          "agent_port": {
            "description": "Agent HTTP 服务端口（默认 9527）",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "auth_token": {
            "type": [
              "string",
              "null"
            ]
          },
          "hostname": {
            "type": "string"
          },
          "ip_address": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "node": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ServerNode"
              }
            ]
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "HeartbeatRequest": {
        "properties": {
          "cpu_usage": {
            "format": "float",
            "type": "number"
          },
          "disk_usage_percent": {
            "format": "float",
            "type": "number"
          },
          "load_one": {
            "format": "float",
            "type": "number"
          },
          "memory_usage_percent": {
            "format": "float",
            "type": "number"
          }
        },
        "required": [
          "cpu_usage",
          "memory_usage_percent",
          "disk_usage_percent",
          "load_one"
        ],
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "type": "object"
      },
      "LoginResponse": {
        "properties": {
          "must_change_password": {
            "description": "是否需强制修改密码（新装面板首次登录）",
            "type": "boolean"
          },
          "refresh_token": {
            "description": "Refresh Token（长过期，默认 24 小时；用于换取新 Access Token）",
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "token": {
            "description": "Access Token（短过期，默认 15 分钟）",
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "refresh_token",
          "username",
          "role",
          "must_change_password"
        ],
        "type": "object"
      },
...
      "ServerNode": {
        "properties": {
          "agent_port": {
            "description": "Agent HTTP 服务端口（面板据此构造 `http://<ip>:<agent_port>` 发起远程调用）",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "auth_token": {
            "description": "Agent 注册时携带的认证令牌",
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "hostname": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "ip_address": {
            "type": "string"
          },
          "last_heartbeat_at": {
            "description": "最近一次心跳时间（Agent 上报）",
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "metrics_json": {
            "description": "最近指标快照 JSON：{cpu_usage, memory_usage_percent, disk_usage_percent, load_one}",
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "hostname",
          "ip_address",
          "status",
          "created_at"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "FlamePanel 面板后端 REST API（OpenAPI 3.x，由 utoipa 编译期生成）",
    "license": {
      "name": ""
    },
    "title": "FlamePanel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "登录成功"
          },
          "401": {
            "description": "凭据错误"
          },
          "429": {
            "description": "尝试过于频繁"
          }
        },
        "summary": "登录",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "refresh",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "刷新成功"
          },
          "401": {
            "description": "令牌无效/过期"
          }
        },
        "summary": "刷新令牌：必须使用 Refresh Token（Bearer），换取新的 Access Token + 轮换 Refresh Token\n剩余寿命不足一半时重新签发 Refresh Token（滑动过期）；否则复用原 Refresh Token\n刷新令牌（Refresh Token 换新 Access Token + 轮换 Refresh Token）",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/nodes/heartbeat/{id}": {
      "post": {
        "operationId": "heartbeat",
        "parameters": [
          {
            "description": "节点 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "心跳记录成功"
          },
          "401": {
            "description": "Agent token 无效"
          }
        },
        "summary": "Agent 心跳上报（白名单免 JWT，校验 Agent token）",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/nodes/register": {
      "post": {
        "operationId": "register_agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "注册成功，返回节点 ID"
          },
          "400": {
            "description": "参数错误"
          }
        },
        "summary": "Agent 注册（白名单免 JWT，供 Agent 启动时自动注册）\n请求体：`{name, host, agent_port, auth_token}`，返回 `{\"id\": <node_id>}`",
        "tags": [
          "nodes"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "默认",
      "url": "/"
    }
  ]
}